Delay loading of video script
