+++ styles/styles.css
@@ -252,3 +252,26 @@ main .section {
 main .section.highlight {
   background-color: var(--highlight-background-color);
 }
+
+/** Video section */
+.vjs-text-track-display, .vjs-control-bar, .vjs-dock-text, .vjs-poster {
+  display: none;
+}
+
+.video-js {
+  position: absolute;
+  top: calc(50% - 174px);
+  left: 0;
+  width: 100%;
+  height: 348px;
+  background-color: transparent;
+  z-index: 10;
+}
+
+@media (min-width: 900px) {
+  .video-container {
+    position: relative;
+    width: 50%;
+    height: auto;
+  }
+}
